Rename files and add kuzzle-sdk to devDependencies
