+++ kepler-companion/package.json
@@ -20,23 +20,25 @@
   },
   "homepage": "https://github.com/kuzzleio/kepler#readme",
   "dependencies": {
-    "axios": "^0.25.0",
     "getmac": "^5.20.0"
   },
   "peerDependencies": {
     "kuzzle-sdk": "^7.0.0"
   },
   "devDependencies": {
-    "@types/node": "^17.0.16",
+    "@types/node": "^17.0.18",
     "@types/node-fetch": "^3.0.3",
     "@typescript-eslint/eslint-plugin": "^5.11.0",
     "@typescript-eslint/parser": "^4.22.0",
     "eslint": "8.8.0",
+    "kuzzle-sdk": "^7.8.4",
     "ts-node": "^9.1.1",
     "typescript": "~4.4.0"
   },
   "files": [
     "index.js",
-    "index.d.ts"
+    "index.d.ts",
+    "KeplerCompanion.js",
+    "KeplerCompanion.d.ts"
   ]
 }
